feat: implement date formatting utility and update related components for consistent date display
